Fix formatting of architectures field in library.properties
The architectures field of library.properties is a comma separated list. The previous incorrect formatting caused the Arduino IDE to display a warning every time the library was compiled:

WARNING: library fp64lib claims to run on (avr atmelavr) architecture(s) and may be incompatible with your current board which runs on (avr) architecture(s).